Refactor ID types from int to Long pour tout les model

+++ src/main/java/ma/cabinetplus/dao/PatientDAO.java
@@ -2,7 +2,7 @@ package ma.cabinetplus.dao;
 
 import ma.cabinetplus.model.Patient;
 
-public interface PatientDAO extends GenericDAO<Patient, Integer> {
+public interface PatientDAO extends GenericDAO<Patient, Long> {
     Patient trouverParNumeroDossier(String numeroDossier);
     Patient trouverParUsername(String username);
 }
