+++ src/main/java/ma/cabinetplus/dao/RendezVousDAO.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface RendezVousDAO extends GenericDAO<RendezVous, Long> {
     void mettreAJourStatut(Long id, StatutRendezVous statut);
-    List<RendezVous> trouverParPatient(int patientId);
+    List<RendezVous> trouverParPatient(Long patientId);
 }
